app/requirements.txt
django
djangorestframework
drf-yasg
django-cors-headers
psycopg2-binary
gunicorn
Pillow